requirements.txt
asgiref==3.5.0
contextlib2==21.6.0
Django==3.2.12
djangorestframework==3.13.1
psycopg2==2.9.3
pytz==2022.1
schema==0.7.5
sqlparse==0.4.2
